+++ packages/react/src/useGlobalData.ts
@@ -1,31 +1,20 @@
-import { useGlobalData as baseUseGlobalData } from '@goldfishjs/composition-api';
-import useContextType from './useContextType';
-import { global } from './Global';
+import { default as Global, global } from './Global';
 
-function check(key: any) {
-  if (!global.data) {
-    throw new Error('The global data is not ready.');
-  }
-
-  if (!(key in global.data)) {
-    throw new Error(`The key: ${key} is not in global data.`);
-  }
-}
-
-export default function useGlobalData<G extends Record<string, any>>() {
-  const type = useContextType();
-  if (type === 'react') {
-    return {
-      get<T extends keyof G>(key: T) {
-        check(key);
-        return global.data[key] as G[T];
-      },
-      set<T extends keyof G>(key: T, value: G[T]) {
-        check(key);
-        global.data[key] = value;
-      },
-    };
-  }
-
-  return baseUseGlobalData<G>();
+export default function useGlobalData<G extends Record<string, any>>(passInGlobal?: Global) {
+  const realGlobal = passInGlobal || global;
+  return {
+    get: <T extends keyof G>(key: T) => {
+      return realGlobal.reactiveData.data[key as string];
+    },
+    set: <T extends keyof G>(key: T, value: G[T]) => {
+      if (key in realGlobal.reactiveData) {
+        realGlobal.reactiveData.data[key as string] = value;
+      } else {
+        realGlobal.reactiveData.data = {
+          ...global.reactiveData.data,
+          [key]: value,
+        };
+      }
+    },
+  };
 }